Remove unsupported 'paths' parameter from publish:github:pull-request action
The publish:github:pull-request action in this RHDH version doesn't support
the 'paths' parameter. All files from sourcePath are automatically included
in the PR, so this parameter is unnecessary.

Fixed templates:
- add-vm-to-namespace/template.yaml
- windows-vm/template.yaml

This maintains the same functionality while resolving the error:
"Invalid input passed to action publish:github:pull-request, instance is not
allowed to have the additional property 'paths'"